+++ NewWorld/src/NewWorld/Math/Functions.h
@@ -6,11 +6,11 @@ namespace NewWorld::Math
 {
 	inline float DegreesToRadians(float degrees)
 	{
-		return degrees / 180 * PI;
+		return degrees / 180.0f * (float)PI;
 	}
 
 	inline float RadiansToDegrees(float radians)
 	{
-		return radians / PI * 180;
+		return radians / (float)PI * 180.0f;
 	}
 }
